proyecto/influencer/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.9
pydantic==2.5.0
pulsar-client==3.7.0
pulsar-client[avro]==3.7.0
alembic==1.13.1
python-multipart==0.0.6
